+++ packages/docs/docs/changelog.md
@@ -1,6 +1,18 @@
 # Changelog
 
-## 0.5.11
+## 0.5.14
+
+- Add [Llama2 support](./guides/models.md#llama2).
+
+## [0.5.13](https://github.com/fixie-ai/ai-jsx/commit/80e25c7d701d0d227e6815f4303ca7dc28dfce0c)
+
+- Add [`DocsQAWithSources` component](./guides/docsqa.md#handling-a-query)
+
+## 0.5.12
+
+- Updated `readme.md` in the `ai-jsx` package to fix bugs on the npm landing page.
+
+## [0.5.11](https://github.com/fixie-ai/ai-jsx/commit/7d5c0fca9c9e1088be7fa0c8a2c74a7db2745e9d)
 
 - Make JIT UI stream rather than appear all at once.
 - Use `openai-edge` instead of `@nick.heiner/openai-edge`